publish cleanup
