+++ Jenkins.NET.Publishing/Jenkins.NET.Publishing.csproj
@@ -81,11 +81,8 @@
     <Reference Include="NuGet.Versioning, Version=4.2.0.0, Culture=neutral, PublicKeyToken=31bf3856ad364e35, processorArchitecture=MSIL">
       <HintPath>..\packages\NuGet.Versioning.4.2.0\lib\net45\NuGet.Versioning.dll</HintPath>
     </Reference>
-    <Reference Include="Photon.Framework, Version=0.0.70.0, Culture=neutral, processorArchitecture=MSIL">
-      <HintPath>..\packages\Photon.Framework.0.0.70\lib\net46\Photon.Framework.dll</HintPath>
-    </Reference>
-    <Reference Include="Photon.MSBuild, Version=0.0.6.0, Culture=neutral, processorArchitecture=MSIL">
-      <HintPath>..\packages\Photon.MSBuild.0.0.6\lib\net46\Photon.MSBuild.dll</HintPath>
+    <Reference Include="Photon.Framework, Version=0.0.74.0, Culture=neutral, processorArchitecture=MSIL">
+      <HintPath>..\packages\Photon.Framework.0.0.74\lib\net46\Photon.Framework.dll</HintPath>
     </Reference>
     <Reference Include="Photon.NuGet, Version=0.0.14.0, Culture=neutral, processorArchitecture=MSIL">
       <HintPath>..\packages\Photon.NuGet.0.0.14\lib\net46\Photon.NuGet.dll</HintPath>
@@ -102,14 +99,16 @@
     <Reference Include="WindowsBase" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Build_Linux.cs" />
-    <Compile Include="Publish_Windows.cs" />
-    <Compile Include="Build_Windows.cs" />
+    <Compile Include="Scripts\Publish.cs" />
+    <Compile Include="Scripts\Build.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="Tools\BuildTools.cs" />
+    <Compile Include="Tools\TestTools.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="app.config" />
     <None Include="packages.config" />
   </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
 </Project>
