+++ Jenkins.NET.Publishing/packages.config
@@ -17,8 +17,7 @@
   <package id="NuGet.Repositories" version="4.2.0" targetFramework="net46" />
   <package id="NuGet.RuntimeModel" version="4.2.0" targetFramework="net46" />
   <package id="NuGet.Versioning" version="4.2.0" targetFramework="net46" />
-  <package id="Photon.Framework" version="0.0.70" targetFramework="net46" />
-  <package id="Photon.MSBuild" version="0.0.6" targetFramework="net46" />
+  <package id="Photon.Framework" version="0.0.74" targetFramework="net46" />
   <package id="Photon.NuGet" version="0.0.14" targetFramework="net46" />
   <package id="Photon.NuGet.Core" version="0.0.3" targetFramework="net46" />
   <package id="Photon.NUnit" version="0.0.1" targetFramework="net46" />
